requirements.txt
streamlit
pymongo
boto3
python-dotenv
